Consolidate remaining arch-specific implementations
