+++ plugins/inputs/procstat/arch_ohers.go
@@ -2,8 +2,18 @@
 
 package procstat
 
-import "github.com/shirou/gopsutil/v3/process"
+import (
+	"errors"
+
+	"github.com/shirou/gopsutil/v3/process"
+)
 
 func processName(p *process.Process) (string, error) {
 	return p.Exe()
 }
+
+func queryPidWithWinServiceName(_ string) (uint32, error) {
+	return 0, errors.New("os not supporting win_service option")
+}
+
+func collectMemmap(Process, string, map[string]any) {}
