--- plugins/inputs/procstat/memmap_notlinux.go
@@ -1,5 +1,0 @@
-//go:build !linux
-
-package procstat
-
-func collectMemmap(Process, string, map[string]any) {}
