--- plugins/inputs/procstat/win_service_notwindows.go
@@ -1,11 +1,0 @@
-//go:build !windows
-
-package procstat
-
-import (
-	"fmt"
-)
-
-func queryPidWithWinServiceName(_ string) (uint32, error) {
-	return 0, fmt.Errorf("os not support win_service option")
-}
